upload success image

+++ eclipse-workspace/HappyGraden/src/main/java/vn/com/unit/repository/ProductImg2DRepository.java
@@ -7,5 +7,5 @@ import vn.com.unit.entity.ProductImg2D;
 
 public interface ProductImg2DRepository extends MirageRepository<ProductImg2D, Long>{
 
-	public void saveImg2D(@Param("id") Long id, @Param("url") String url);
+	public void saveImg2D(@Param("id") int id, @Param("url") String url);
 }
